# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rhizom API",
    "description": "REST API of a Rhizom server. Note paths are vault paths: POSIX, relative to the vault root, with extension.",
    "version": "0.0.0"
  },
...
  "paths": {
...
    "/api/vault": {
      "get": {
        "summary": "Describe the open vault",
        "tags": ["vault"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["name", "noteCount", "indexedAt"],
                  "properties": {
                    "name": {
                      "description": "Folder name of the vault",
                      "type": "string"
                    },
                    "noteCount": {
                      "type": "integer"
                    },
                    "indexedAt": {
                      "anyOf": [
                        {
                          "format": "date-time",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tree": {
      "get": {
        "summary": "Folder tree of the vault",
        "tags": ["notes"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "type": "object",
                        "required": ["type", "name", "path", "children"],
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": ["folder"]
                          },
                          "name": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          },
                          "children": {
                            "type": "array",
                            "items": {
                              "$ref": "#/components/schemas/def-0"
                            }
                          }
                        }
                      },
                      {
                        "type": "object",
                        "required": ["type", "name", "path", "title"],
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": ["note"]
                          },
                          "name": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          }
                        }
                      }
                    ],
                    "title": "TreeEntry"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notes": {
      "get": {
        "summary": "List all notes",
        "tags": ["notes"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "path",
                      "name",
                      "title",
                      "folder",
                      "tags",
                      "modifiedAt",
                      "size",
                      "linkCount",
                      "backlinkCount"
                    ],
                    "properties": {
                      "path": {
                        "description": "Vault path of the note, including the extension",
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "folder": {
                        "type": "string"
                      },
                      "tags": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "modifiedAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "size": {
                        "type": "integer"
                      },
                      "linkCount": {
                        "type": "integer"
                      },
                      "backlinkCount": {
                        "type": "integer"
                      }
                    },
                    "title": "NoteSummary"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a note",
        "tags": ["notes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["path"],
                "properties": {
                  "path": {
                    "minLength": 1,
                    "description": "Vault path; .md is added when missing",
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "path",
                    "name",
                    "title",
                    "folder",
                    "tags",
                    "modifiedAt",
                    "size",
                    "linkCount",
                    "backlinkCount",
                    "content",
                    "hash",
                    "frontmatter",
                    "headings"
                  ],
                  "properties": {
                    "path": {
                      "description": "Vault path of the note, including the extension",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "folder": {
                      "type": "string"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "modifiedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "size": {
                      "type": "integer"
                    },
                    "linkCount": {
                      "type": "integer"
                    },
                    "backlinkCount": {
                      "type": "integer"
                    },
                    "content": {
                      "type": "string"
                    },
                    "hash": {
                      "description": "Send back as If-Match when saving",
                      "type": "string"
                    },
                    "frontmatter": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "headings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["level", "text", "slug", "line"],
                        "properties": {
                          "level": {
                            "minimum": 1,
                            "maximum": 6,
                            "type": "integer"
                          },
                          "text": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "line": {
                            "minimum": 0,
                            "type": "integer"
                          }
                        },
                        "title": "Heading"
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/links": {
      "get": {
        "summary": "Outgoing links of a note",
        "tags": ["notes"],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "query",
            "name": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["source", "target", "raw", "kind", "line"],
                    "properties": {
                      "source": {
                        "type": "string"
                      },
                      "target": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "raw": {
                        "type": "string"
                      },
                      "kind": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": ["wikilink"]
                          },
                          {
                            "type": "string",
                            "enum": ["embed"]
                          },
                          {
                            "type": "string",
                            "enum": ["markdown"]
                          }
                        ]
                      },
                      "alias": {
                        "type": "string"
                      },
                      "heading": {
                        "type": "string"
                      },
                      "line": {
                        "type": "integer"
                      }
                    },
                    "title": "NoteLink"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/backlinks": {
      "get": {
        "summary": "Notes that link to a note",
        "tags": ["notes"],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "query",
            "name": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["source", "sourceTitle", "context", "line"],
                    "properties": {
                      "source": {
                        "type": "string"
                      },
                      "sourceTitle": {
                        "type": "string"
                      },
                      "context": {
                        "type": "string"
                      },
                      "line": {
                        "type": "integer"
                      }
                    },
                    "title": "Backlink"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Shape of every error response",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Shape of every error response",
                  "type": "object",
                  "required": ["statusCode", "error", "message"],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}